refactor(server): 简化 OAuth 工具调用流程
将 OAuth 工具的特殊处理逻辑从 server.ts 移至 app/handlers.ts：
- 新增 oauthState.ts：统一的 OAuth 状态管理器
- OAuth 工具现在走统一的 tool call 流程，无需特殊判断
- 移除 server.ts 中重复的 handleOAuthStart/handleOAuthCompletion 方法
- 代码更简洁，职责更清晰

优化：
- server.ts 减少约 70 行代码
- 所有工具统一处理，无特殊逻辑
- OAuth 状态管理更集中、更易维护
