+++ src/features/app/handlers.ts
@@ -8,6 +8,10 @@ import { getAllDevelopersAndApps, selectApp as selectAppApi } from './api.js';
 import { clearAppCache } from '../../core/utils/cache.js';
 import { clearToken } from '../../core/auth/tokenStorage.js';
 import { ApiConfig } from '../../core/network/httpClient.js';
+import { requestDeviceCode, generateAuthUrl, pollForToken } from '../../core/auth/oauth.js';
+import { saveToken } from '../../core/auth/tokenStorage.js';
+import { oauthState } from '../../core/auth/oauthState.js';
+import { getMacTokenStatus, getTokenSourceLabel } from '../../core/utils/handlerHelpers.js';
 
 /**
  * List all developers and apps for the current user
@@ -142,6 +146,79 @@ ${error instanceof Error ? error.message : String(error)}
 
 请使用 \`check_environment\` 工具验证您的配置。
 `;
+  }
+}
+
+/**
+ * Start OAuth authorization
+ */
+export async function startOAuthAuthorization(context: HandlerContext): Promise<string> {
+  // Use shared authentication check logic
+  const { hasMacToken, source } = getMacTokenStatus(context);
+
+  if (hasMacToken) {
+    const sourceLabel = getTokenSourceLabel(source);
+    return '✅ 已经完成授权\n\n' +
+           `当前已有有效的 MAC Token ${sourceLabel}，可以直接使用所有功能。\n\n` +
+           '💡 如需切换账号，请先使用 clear_auth_data 工具清除现有授权。';
+  }
+
+  try {
+    const apiConfig = ApiConfig.getInstance();
+    const environment = apiConfig.environment;
+    const deviceCodeData = await requestDeviceCode(environment);
+    const authUrl = generateAuthUrl(deviceCodeData.qrcode_url, environment);
+    
+    // 保存状态，供 completion 使用
+    oauthState.setPendingState({
+      deviceCode: deviceCodeData.device_code,
+      environment
+    });
+
+    return '🔐 TapTap 授权登录\n\n' +
+           '请按以下步骤完成授权：\n\n' +
+           `1️⃣ 打开授权链接：\n   ${authUrl}\n\n` +
+           '2️⃣ 使用 TapTap App 扫描二维码\n\n' +
+           '3️⃣ 授权成功后，调用 complete_oauth_authorization 工具完成授权\n\n' +
+           '💡 提示：授权链接有效期为 2 分钟，过期后需要重新获取';
+  } catch (error) {
+    return `❌ 获取授权链接失败: ${error instanceof Error ? error.message : String(error)}\n\n` +
+           '请稍后重试或联系技术支持。';
+  }
+}
+
+/**
+ * Complete OAuth authorization
+ */
+export async function completeOAuthAuthorization(): Promise<string> {
+  const pendingState = oauthState.getPendingState();
+  
+  if (!pendingState) {
+    return '❌ 未找到待完成的授权\n\n' +
+           '请先使用 start_oauth_authorization 工具获取授权链接。';
+  }
+
+  try {
+    const macToken = await pollForToken(pendingState.deviceCode, pendingState.environment);
+    
+    // 保存 token
+    saveToken(macToken, { environment: pendingState.environment });
+    const apiConfig = ApiConfig.getInstance();
+    apiConfig.setMacToken(macToken);
+    
+    // 清除状态
+    oauthState.clearPendingState();
+
+    return '✅ 授权完成！\n\n' +
+           'Token 已成功保存，现在可以使用所有需要认证的功能了。\n\n' +
+           '请重新执行之前失败的操作。';
+  } catch (error) {
+    return `❌ 授权失败: ${error instanceof Error ? error.message : String(error)}\n\n` +
+           '请确认：\n' +
+           '1. 已在浏览器中打开授权链接\n' +
+           '2. 已使用 TapTap App 扫码授权\n' +
+           '3. 授权页面显示成功\n\n' +
+           '如果仍然失败，请使用 start_oauth_authorization 工具获取新的授权链接。';
   }
 }
